+++ child2.cpp
@@ -1,12 +1,10 @@
 #include <iostream>
 #include <fstream>
 #include <unistd.h>
-#include <cstring>
-#include <sys/wait.h>
 #include <signal.h>
-#include <wait.h>
 
 using namespace std;
+
 bool flag = true;
 
 void signal_handler(int signum) {
@@ -22,22 +20,27 @@ int main(int argc, char** argv) {
 	cout << "P2 start" << endl;
 	sigset_t mask;
 	sigemptyset(&mask);
-	sigaddset(&mask, SIGUSR2);
-	sigprocmask(SIG_BLOCK, &mask, nullptr);
-
-	signal(SIGINT, signal_handler);
+    sigaddset(&mask, SIGUSR1);
+    sigaddset(&mask, SIGUSR2);
+    sigprocmask(SIG_SETMASK, &mask, NULL);
+	sigemptyset(&mask);
+    sigaddset(&mask, SIGUSR2);
+	signal(SIGQUIT, signal_handler);
 	
 	sigwait(&mask, &signum);
-	while((bytes = read(*argv[2], &c, 1)) || flag) {
-		// if (bytes == 0)
-		// 	break;
-		cout << "P2 ch: " << c << endl;
-		out << c;
-		kill(0, SIGUSR1);
-		sigwait(&mask, &signum);		
+	//while((bytes = read(*argv[2], &c, 1)) || flag) {
+	while(true) {
+		if (read(*argv[2], &c, 1)) {
+			cout << "P2 ch: " << c << endl;
+			out << c;
+			kill(0, SIGUSR1);
+			sigwait(&mask, &signum);
+		} else {
+			if (flag == false) break;
+		}	
 	}
 	kill(0, SIGUSR1);
 	out.close();
-	exit(EXIT_SUCCESS);
+
 	return 0;
 }
